Add Fragment lifecycle and several sample cleanups
This adds Fragment Lifecycle and several sample cleanups. The path
forward is to keep iterating on the Sample for it to become a more
complex example. I will expand it to have more features than the
current PR listing. Will make an intermediary interaction with details
and navigation of each Repository.